requirements.txt
django
djangorestframework
psycopg2-binary
django-environ
celery
redis
mysqlclient
Pillow
django-storages
djangorestframework
pytest 
pytest-django
djangorestframework-simplejwt
django-redis 
